CT-184/Modify : data type error

+++ src/main/java/com/xeat/llmservice/Controller/LLMController.java
@@ -37,14 +37,14 @@ public class LLMController {
 
     @Operation(summary = "코딩테스트 최근 채팅 조회 API", description = "코딩테스트 가장 최근 채팅 및 채팅 정보를 조회하는 API")
     @GetMapping("/{codeHistoryId}")
-    public ResponseEntity<LLMResponseDTO.ChatResponseList> chatHistory(@PathVariable Integer codeHistoryId,
+    public ResponseEntity<LLMResponseDTO.ChatResponseList> chatHistory(@PathVariable Long codeHistoryId,
                                                                    @RequestHeader("UserId") String userId) {
         return llmService.chatHistory(userId, codeHistoryId);
     }
 
     @Operation(summary = "코딩테스트 채팅 페이지네이션 조회 API", description = "코딩테스트 채팅을 조회하는 API")
     @GetMapping("/history/{codeHistoryId}")
-    public ResponseEntity<LLMResponseDTO.ChatResponseList> chatPagedHistory(@PathVariable Integer codeHistoryId, @RequestParam("page") Integer page,
+    public ResponseEntity<LLMResponseDTO.ChatResponseList> chatPagedHistory(@PathVariable Long codeHistoryId, @RequestParam("page") Integer page,
                                                                    @RequestHeader("UserId") String userId) {
         return llmService.chatPagedHistory(userId,codeHistoryId, page);
     }
