+++ src/main/java/com/xeat/llmservice/Repository/LLMRepository.java
@@ -6,8 +6,8 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface LLMRepository extends JpaRepository <LLMEntity, Long> {
-    boolean existsByCodeHistoryId(Integer codeHistoryId);
-    LLMEntity findByCodeHistoryId(Integer codeHistoryId);
+    boolean existsByCodeHistoryId(Long codeHistoryId);
+    LLMEntity findByCodeHistoryId(Long codeHistoryId);
 
     boolean existsByUserId(String userId);
 }
